+++ nbproject/private/private.properties
@@ -12,6 +12,6 @@ user.properties.file=C:\\Users\\mrtkb\\AppData\\Roaming\\NetBeans\\22\\build.pro
 <<<<<<< Updated upstream
 user.properties.file=C:\\Users\\mrtkb\\AppData\\Roaming\\NetBeans\\22\\build.properties
 =======
-user.properties.file=C:\\Users\\susanyaaaa\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
+user.properties.file=C:\\Users\\CHAMA COMPUTERS\\AppData\\Roaming\\NetBeans\\22\\build.properties
 >>>>>>> Stashed changes
 >>>>>>> 122b1f24780513645bf4e7cf7f73d7dc3ec52f14
