--- src/gui/AllCourses1.java
@@ -1,265 +1,0 @@
-
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
- */
-package gui;
-
-import com.formdev.flatlaf.FlatDarkLaf;
-import java.sql.ResultSet;
-import java.util.Vector;
-import javax.swing.SwingConstants;
-import javax.swing.table.DefaultTableCellRenderer;
-import javax.swing.table.DefaultTableModel;
-import model.MySQL2;
-
-/**
- *
- * @author mrtkb
- */
-public class AllCourses1 extends javax.swing.JFrame {
-
-    private StudentPayment studentPayment;
-
-    public void setInvoice(StudentPayment studentPayment) {
-        this.studentPayment = studentPayment;
-    }
-
-    /**
-     * Creates new form allSubjects
-     */
-    public AllCourses1() {
-        initComponents();
-
-        DefaultTableCellRenderer renderer = new DefaultTableCellRenderer();
-        renderer.setHorizontalAlignment(SwingConstants.CENTER);
-
-        jTable1.setDefaultRenderer(Object.class, renderer);
-
-        LoadSubject();
-
-    }
-
-    private void LoadSubject() {
-        
-        try {
-            
-            String sort = String.valueOf(jComboBox1.getSelectedItem());
-            
-            String searchText = jTextField1.getText().toLowerCase();
-            
-            String query = "SELECT * FROM `courses`";
-            
-            if (!searchText.isEmpty()) {
-                
-                query += "WHERE LOWER(`id`) LIKE '%" + searchText + "%' "
-                        + "OR LOWER(`name`) LIKE '%" + searchText + "%' "
-                        + "OR LOWER(`fee`) LIKE '%" + searchText + "%' ";
-                
-            }
-            
-            if (sort.equals("Price ASC")) {
-                query += "ORDER BY `fee` ASC";
-            } else if (sort.equals("Price DESC")) {
-                query += "ORDER BY `fee` DESC";
-            } else if (sort.equals("Name ASC")) {
-                query += "ORDER BY `name` ASC";
-            } else if (sort.equals("Name DESC")) {
-                query += "ORDER BY `name` DESC";
-            } else if (sort.equals("ID ASC")) {
-                query += "ORDER BY `id` ASC";
-            } else if (sort.equals("ID DESC")) {
-                query += "ORDER BY `id` DESC";
-            }
-            
-            // Execute SQL query to retrieve subjects
-            ResultSet rs = MySQL2.executeSearch(query);
-
-            // Get the table model for jTable1
-            DefaultTableModel model = (DefaultTableModel) jTable1.getModel();
-            model.setRowCount(0); // Clear existing rows
-
-            // Iterate through the result set and populate jTable1
-            while (rs.next()) {
-                Vector<Object> vectorE = new Vector<>();
-                vectorE.add(rs.getInt("id"));
-                vectorE.add(rs.getString("name"));
-                vectorE.add(rs.getString("fee"));
-
-                model.addRow(vectorE); // Add row to the table model
-            }
-
-            jTable1.setModel(model); // Set the updated model to jTable1
-//        LOGGER.info("Subject data successfully loaded.");
-        } catch (Exception ex) {
-            ex.printStackTrace();
-//        LOGGER.log(Level.SEVERE, "Error loading subject data", ex);
-        }
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        jScrollPane1 = new javax.swing.JScrollPane();
-        jTable1 = new javax.swing.JTable();
-        jLabel1 = new javax.swing.JLabel();
-        jLabel2 = new javax.swing.JLabel();
-        jComboBox1 = new javax.swing.JComboBox<>();
-        jLabel3 = new javax.swing.JLabel();
-        jTextField1 = new javax.swing.JTextField();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-
-        jTable1.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][] {
-
-            },
-            new String [] {
-                "Id", "Name", "Price"
-            }
-        ) {
-            boolean[] canEdit = new boolean [] {
-                false, false, false
-            };
-
-            public boolean isCellEditable(int rowIndex, int columnIndex) {
-                return canEdit [columnIndex];
-            }
-        });
-        jTable1.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                jTable1MouseClicked(evt);
-            }
-        });
-        jScrollPane1.setViewportView(jTable1);
-
-        jLabel1.setFont(new java.awt.Font("Century Gothic", 0, 36)); // NOI18N
-        jLabel1.setText("All Courses");
-
-        jLabel2.setFont(new java.awt.Font("Century Gothic", 1, 14)); // NOI18N
-        jLabel2.setText("Sort By :");
-
-        jComboBox1.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Price ASC", "Price DESC", "Name ASC", "Name DESC", "ID ASC", "ID DESC" }));
-        jComboBox1.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jComboBox1ActionPerformed(evt);
-            }
-        });
-
-        jLabel3.setFont(new java.awt.Font("Century Gothic", 1, 14)); // NOI18N
-        jLabel3.setText("Search");
-
-        jTextField1.addKeyListener(new java.awt.event.KeyAdapter() {
-            public void keyReleased(java.awt.event.KeyEvent evt) {
-                jTextField1KeyReleased(evt);
-            }
-        });
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addGap(406, 406, 406)
-                .addComponent(jLabel1)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-            .addGroup(layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 988, Short.MAX_VALUE)
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                        .addComponent(jLabel2)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addComponent(jComboBox1, javax.swing.GroupLayout.PREFERRED_SIZE, 132, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addComponent(jLabel3)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, 204, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap())
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                .addGap(10, 10, 10)
-                .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 39, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 39, Short.MAX_VALUE)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                        .addComponent(jComboBox1, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addComponent(jLabel2))
-                    .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                        .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addComponent(jLabel3, javax.swing.GroupLayout.PREFERRED_SIZE, 21, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 357, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap())
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void jTable1MouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jTable1MouseClicked
-        int row = jTable1.getSelectedRow();
-
-        if (evt.getClickCount() == 2) {
-    // LOGGER.info("Table row double-clicked, selected customer: " + jTable1.getValueAt(row, 0));
-    if (studentPayment != null) {
-        studentPayment.getSubjectNameField().setText(String.valueOf(jTable1.getValueAt(row, 1)));
-        
-        try {
-            // Convert column 0 (subject ID) to an integer
-            int subjectId = Integer.parseInt(jTable1.getValueAt(row, 0).toString());
-            // Convert column 2 (selling price) to a double
-            double selling_p = Double.parseDouble(jTable1.getValueAt(row, 2).toString());
-
-            studentPayment.setSubject_id(subjectId);
-            studentPayment.setSelling_price(selling_p);
-
-            this.dispose();
-        } catch (NumberFormatException e) {
-            // Handle the exception if the data is not in the expected format
-            System.err.println("Error parsing number: " + e.getMessage());
-        }
-    }
-}
-
-    }//GEN-LAST:event_jTable1MouseClicked
-
-    private void jComboBox1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jComboBox1ActionPerformed
-        LoadSubject();
-    }//GEN-LAST:event_jComboBox1ActionPerformed
-
-    private void jTextField1KeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jTextField1KeyReleased
-        LoadSubject();
-    }//GEN-LAST:event_jTextField1KeyReleased
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-
-        FlatDarkLaf.setup();
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new AllCourses1().setVisible(true);
-            }
-        });
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JComboBox<String> jComboBox1;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JLabel jLabel2;
-    private javax.swing.JLabel jLabel3;
-    private javax.swing.JScrollPane jScrollPane1;
-    private javax.swing.JTable jTable1;
-    private javax.swing.JTextField jTextField1;
-    // End of variables declaration//GEN-END:variables
-}
