+++ src/gui/AllStudents.java
@@ -38,8 +38,23 @@ public class AllStudents extends javax.swing.JPanel {
         try {
 
             String sort = String.valueOf(jComboBox1.getSelectedItem());
+            
+            String searchText = jTextField1.getText().toLowerCase();
 
-            String query = "SELECT * FROM `student` INNER JOIN `gender` ON `student`.`gender_id` = `gender`.`id`";
+            String query = "SELECT `nic`, `first_name`, `last_name`, `dob`, `contact_info`, `registration_date`, `email`, `gender`.`name` FROM `student` "
+                    + "INNER JOIN `gender` ON `student`.`gender_id` = `gender`.`id`";
+            
+            if (!searchText.isEmpty()) {
+
+                query += "WHERE (LOWER(`nic`) LIKE '%" + searchText + "%' "
+                        + "OR LOWER(`first_name`) LIKE '%" + searchText + "%' "
+                        + "OR LOWER(`last_name`) LIKE '%" + searchText + "%' "
+                        + "OR LOWER(`dob`) LIKE '%" + searchText + "%' "
+                        + "OR LOWER(`contact_info`) LIKE '%" + searchText + "%' "
+                        + "OR LOWER(`registration_date`) LIKE '%" + searchText + "%' "
+                        + "OR LOWER(`email`) LIKE '%" + searchText + "%') ";
+
+            }
 
             if (sort.equals("First Name ASC")) {
                 query += "ORDER BY `student`.`first_name` ASC";
@@ -90,7 +105,7 @@ public class AllStudents extends javax.swing.JPanel {
         jComboBox1 = new javax.swing.JComboBox<>();
         jButton1 = new javax.swing.JButton();
         jLabel3 = new javax.swing.JLabel();
-        searchName = new javax.swing.JTextField();
+        jTextField1 = new javax.swing.JTextField();
 
         jLabel1.setFont(new java.awt.Font("Century Gothic", 1, 24)); // NOI18N
         jLabel1.setText("All Students");
@@ -132,14 +147,14 @@ public class AllStudents extends javax.swing.JPanel {
         });
 
         jButton1.setFont(new java.awt.Font("Century Gothic", 1, 14)); // NOI18N
-        jButton1.setText("Generate Report");
+        jButton1.setText("Print");
 
         jLabel3.setFont(new java.awt.Font("Century Gothic", 1, 14)); // NOI18N
-        jLabel3.setText("Search By Name or NIC:");
+        jLabel3.setText("Search");
 
-        searchName.addKeyListener(new java.awt.event.KeyAdapter() {
+        jTextField1.addKeyListener(new java.awt.event.KeyAdapter() {
             public void keyReleased(java.awt.event.KeyEvent evt) {
-                searchNameKeyReleased(evt);
+                jTextField1KeyReleased(evt);
             }
         });
 
@@ -148,22 +163,24 @@ public class AllStudents extends javax.swing.JPanel {
         layout.setHorizontalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
-                .addGap(25, 25, 25)
-                .addComponent(jLabel2)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jComboBox1, javax.swing.GroupLayout.PREFERRED_SIZE, 128, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(18, 18, 18)
-                .addComponent(jLabel3)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(searchName, javax.swing.GroupLayout.PREFERRED_SIZE, 145, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(jButton1)
-                .addGap(23, 23, 23))
-            .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 1000, Short.MAX_VALUE)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(jLabel1)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap()
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jScrollPane1)
+                    .addGroup(layout.createSequentialGroup()
+                        .addComponent(jLabel2)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                        .addComponent(jComboBox1, javax.swing.GroupLayout.PREFERRED_SIZE, 128, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 472, Short.MAX_VALUE)
+                        .addComponent(jLabel3)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                        .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, 182, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jButton1))
+                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
+                        .addGap(0, 0, Short.MAX_VALUE)
+                        .addComponent(jLabel1)
+                        .addGap(0, 0, Short.MAX_VALUE)))
+                .addContainerGap())
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -176,7 +193,7 @@ public class AllStudents extends javax.swing.JPanel {
                     .addComponent(jComboBox1, javax.swing.GroupLayout.PREFERRED_SIZE, 35, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addComponent(jButton1, javax.swing.GroupLayout.PREFERRED_SIZE, 35, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addComponent(jLabel3)
-                    .addComponent(searchName, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, 36, javax.swing.GroupLayout.PREFERRED_SIZE))
                 .addGap(28, 28, 28)
                 .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(0, 0, 0))
@@ -241,106 +258,13 @@ public class AllStudents extends javax.swing.JPanel {
         }
     }//GEN-LAST:event_jTable1MouseClicked
 
-    private void searchNameKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_searchNameKeyReleased
-        String searchText = searchName.getText().trim();
-        loadTableWithSearch(searchText);
-    }//GEN-LAST:event_searchNameKeyReleased
+    private void jTextField1KeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jTextField1KeyReleased
+        loadTable();
+    }//GEN-LAST:event_jTextField1KeyReleased
 
     private void jComboBox1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jComboBox1ActionPerformed
-        loadSortTable();
+        loadTable();
     }//GEN-LAST:event_jComboBox1ActionPerformed
-
-    private void loadSortTable() {
-        try {
-            String sort = String.valueOf(jComboBox1.getSelectedItem());
-
-            String query = "SELECT * FROM `student` INNER JOIN `gender` ON `student`.`gender_id` = `gender`.`id` ";
-
-            if (sort.equals("First Name ASC")) {
-                query += "ORDER BY `student`.`first_name` ASC";
-            } else if (sort.equals("First Name DESC")) {
-                query += "ORDER BY `student`.`first_name` DESC";
-            } else if (sort.equals("Last Name ASC")) {
-                query += "ORDER BY `student`.`last_name` ASC";
-            } else if (sort.equals("Last Name DESC")) {
-                query += "ORDER BY `student`.`last_name` DESC";
-            } else if (sort.equals("NIC ASC")) {
-                query += "ORDER BY `student`.`nic` ASC";
-            } else if (sort.equals("NIC DESC")) {
-                query += "ORDER BY `student`.`nic` DESC";
-            }
-
-            ResultSet resultSet = MySQL2.executeSearch(query);
-
-            DefaultTableModel model = (DefaultTableModel) jTable1.getModel();
-            model.setRowCount(0); // Clear the table
-
-            while (resultSet.next()) {
-                Vector<String> vector = new Vector<>();
-                vector.add(resultSet.getString("NIC"));
-                vector.add(resultSet.getString("first_name"));
-                vector.add(resultSet.getString("last_name"));
-                vector.add(resultSet.getString("dob"));
-                vector.add(resultSet.getString("contact_info"));
-                vector.add(resultSet.getString("registration_date"));
-                vector.add(resultSet.getString("email"));
-                vector.add(resultSet.getString("gender.name"));
-
-                model.addRow(vector);
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    private void loadTableWithSearch(String searchText) {
-        try {
-            String sort = String.valueOf(jComboBox1.getSelectedItem());
-            String query = "SELECT * FROM `student` INNER JOIN `gender` ON `student`.`gender_id` = `gender`.`id`";
-
-            if (!searchText.isEmpty()) {
-                query += " WHERE `student`.`first_name` LIKE '%" + searchText + "%' "
-                        + "OR `student`.`last_name` LIKE '%" + searchText + "%' "
-                        + "OR `student`.`nic` LIKE '%" + searchText + "%'";
-            }
-
-            if (sort.equals("First Name ASC")) {
-                query += " ORDER BY `student`.`first_name` ASC";
-            } else if (sort.equals("First Name DESC")) {
-                query += " ORDER BY `student`.`first_name` DESC";
-            } else if (sort.equals("Last Name ASC")) {
-                query += " ORDER BY `student`.`last_name` ASC";
-            } else if (sort.equals("Last Name DESC")) {
-                query += " ORDER BY `student`.`last_name` DESC";
-            } else if (sort.equals("NIC ASC")) {
-                query += " ORDER BY `student`.`nic` ASC";
-            } else if (sort.equals("NIC DESC")) {
-                query += " ORDER BY `student`.`nic` DESC";
-            }
-
-            ResultSet resultSet = MySQL2.executeSearch(query);
-
-            DefaultTableModel model = (DefaultTableModel) jTable1.getModel();
-            model.setRowCount(0);
-
-            while (resultSet.next()) {
-                Vector<String> vector = new Vector<>();
-                vector.add(resultSet.getString("nic"));
-                vector.add(resultSet.getString("first_name"));
-                vector.add(resultSet.getString("last_name"));
-                vector.add(resultSet.getString("dob"));
-                vector.add(resultSet.getString("contact_info"));
-                vector.add(resultSet.getString("registration_date"));
-                vector.add(resultSet.getString("email"));
-                vector.add(resultSet.getString("gender.name"));
-
-                model.addRow(vector);
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton jButton1;
@@ -350,6 +274,6 @@ public class AllStudents extends javax.swing.JPanel {
     private javax.swing.JLabel jLabel3;
     private javax.swing.JScrollPane jScrollPane1;
     private javax.swing.JTable jTable1;
-    private javax.swing.JTextField searchName;
+    private javax.swing.JTextField jTextField1;
     // End of variables declaration//GEN-END:variables
 }
